Add tests for class Plugin
Simply test the 2 methods,
while extracting set_license_validity() into
the helper trait.
