+++ tests/bootstrap.php
@@ -25,7 +25,7 @@ require_once $_tests_dir . '/includes/functions.php';
  */
 function _manually_load_plugin() {
 	require dirname( dirname( __FILE__ ) ) . '/block-lab.php';
-	require __DIR__ . '/php/blocks/controls/trait-control-helper.php';
+	require __DIR__ . '/php/trait-helper.php';
 }
 tests_add_filter( 'muplugins_loaded', '_manually_load_plugin' );
